requirements.txt
facenet-pytorch==2.5.3
opencv-python==4.8.1.78
numpy==1.24.4
torch>=1.9.0
torchvision>=0.10.0
Pillow>=9.0.0
scikit-learn>=1.0.0
